Splitted unique role and group reason page into twice
fixed admin pending page by adding missing reason column
