--- voms-admin-server/src/main/java/org/glite/security/voms/admin/view/actions/user/LoadRoleGroupAction.java
@@ -1,59 +1,0 @@
-/**
- * Copyright (c) Members of the EGEE Collaboration. 2006-2009.
- * See http://www.eu-egee.org/partners/ for details on the copyright holders.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Authors:
- * 	Andrea Ceccanti (INFN)
- */
-
-package org.glite.security.voms.admin.view.actions.user;
-
-import org.apache.struts2.convention.annotation.Result;
-import org.apache.struts2.convention.annotation.Results;
-import org.glite.security.voms.admin.view.actions.BaseAction;
-
-@Results({ @Result(name = BaseAction.SUCCESS, location = "requestGroupRole") })
-public class LoadRoleGroupAction extends UserActionSupport {
-
-  /**
-	 * 
-	 */
-  private static final long serialVersionUID = 1L;
-
-  Long roleId = -1L;
-  Long groupId = -1L;
-
-  public Long getRoleId() {
-    return roleId;
-  }
-
-  public void setRoleId(Long roleId) {
-    this.roleId = roleId;
-  }
-
-  public Long getGroupId() {
-    return groupId;
-  }
-
-  public void setGroupId(Long groupId) {
-    this.groupId = groupId;
-  }
-
-  @Override
-  public String execute() throws Exception {
-
-    return SUCCESS;
-  }
-}
